refactor: reorganize components and layout
